added sepearte sensors class in statemachine which will update snsors_raw struct which will be acsessed directyl by estimator class. Thinking of creating a setup state and setting up all io in there and making vspi an object in the state machine instead of global

+++ Ricardo_OS/Ricardo_OS/src/Comms/downlink.cpp
@@ -12,7 +12,7 @@ void Downlink::setup(){
     _sm->estimator.update();// example of acsessing object called in statemachine
 };
 
-void Downlink::send_data(uint8_t iface,uint32_t* data, size_t len){
+void Downlink::send_data(uint8_t iface,uint8_t* data, size_t len){
     
 
 };
