+++ Ricardo_OS/Ricardo_OS/src/Comms/interfaces/iface.h
@@ -1,4 +1,4 @@
-//parent class of all interfaces
+//abstract class of all interfaces
 #ifndef IFACE_H
 #define IFACE_H
 #include <Arduino.h>
@@ -6,8 +6,8 @@
 class Iface{
     public:
         virtual void setup(int8_t address);
-        virtual void send_packet(uint32_t* txpacket, size_t packet_len);
-        virtual void recieve_packet(uint32_t* rxpacket);
+        virtual void send_packet(uint8_t* txpacket, size_t packet_len);
+        virtual void recieve_packet(uint8_t* rxpacket);
         int8_t address;
